+++ packages/dataviews/src/components/dataviews-filters/filter.tsx
@@ -309,9 +309,9 @@ const FilterText = ( {
 
 		return createInterpolateElement(
 			sprintf(
-				/* translators: 1: Filter name. 2: Min value. 3: Max value. e.g.: "Item count between (inc): 10-180". */
+				/* translators: 1: Filter name. 2: Min value. 3: Max value. e.g.: "Item count between (inc): 10 and 180". */
 				__(
-					'<Name>%1$s between (inc): </Name><Value>%2$s-%3$s</Value>'
+					'<Name>%1$s between (inc): </Name><Value>%2$s and %3$s</Value>'
 				),
 				filter.name,
 				label[ 0 ],