+++ packages/dataviews/CHANGELOG.md
@@ -14,6 +14,7 @@
 ### Features
 
 - Support Ctrl + Click / Cmd + Click for multiselecting rows in the Table layout ([#70891](https://github.com/WordPress/gutenberg/pull/70891)).
+- Add support for the `Edit` control on the date field type ([#70836](https://github.com/WordPress/gutenberg/pull/70836)).
 
 ## 5.0.0 (2025-07-23)